SSKW-3 Adapted useMonthDays to use a utilitary function and removed mock
