--- src/client/__mocks__/hooks/useMonthDaysMock.js
@@ -1,74 +1,0 @@
-module.exports = [
-  {
-    days: [
-      { value: 1587877200000, isOtherMonth: true },
-      { value: 1587963600000, isOtherMonth: true },
-      { value: 1588050000000, isOtherMonth: true },
-      { value: 1588136400000, isOtherMonth: true },
-      { value: 1588222800000, isOtherMonth: true },
-      { value: 1588309200000, isOtherMonth: false },
-      { value: 1588395600000, isOtherMonth: false },
-    ],
-    id: 0,
-  },
-  {
-    days: [
-      { value: 1588482000000, isOtherMonth: false },
-      { value: 1588568400000, isOtherMonth: false },
-      { value: 1588654800000, isOtherMonth: false },
-      { value: 1588741200000, isOtherMonth: false },
-      { value: 1588827600000, isOtherMonth: false },
-      { value: 1588914000000, isOtherMonth: false },
-      { value: 1589000400000, isOtherMonth: false },
-    ],
-    id: 1,
-  },
-  {
-    days: [
-      { value: 1589086800000, isOtherMonth: false },
-      { value: 1589173200000, isOtherMonth: false },
-      { value: 1589259600000, isOtherMonth: false },
-      { value: 1589346000000, isOtherMonth: false },
-      { value: 1589432400000, isOtherMonth: false },
-      { value: 1589518800000, isOtherMonth: false },
-      { value: 1589605200000, isOtherMonth: false },
-    ],
-    id: 2,
-  },
-  {
-    days: [
-      { value: 1589691600000, isOtherMonth: false },
-      { value: 1589778000000, isOtherMonth: false },
-      { value: 1589864400000, isOtherMonth: false },
-      { value: 1589950800000, isOtherMonth: false },
-      { value: 1590037200000, isOtherMonth: false },
-      { value: 1590123600000, isOtherMonth: false },
-      { value: 1590210000000, isOtherMonth: false },
-    ],
-    id: 3,
-  },
-  {
-    days: [
-      { value: 1590296400000, isOtherMonth: false },
-      { value: 1590382800000, isOtherMonth: false },
-      { value: 1590469200000, isOtherMonth: false },
-      { value: 1590555600000, isOtherMonth: false },
-      { value: 1590642000000, isOtherMonth: false },
-      { value: 1590728400000, isOtherMonth: false },
-      { value: 1590814800000, isOtherMonth: false },
-    ],
-    id: 4,
-  },
-  {
-    days: [
-      { value: 1590901200000, isOtherMonth: false },
-      { value: 1590987600000, isOtherMonth: true },
-      { value: 1591074000000, isOtherMonth: true },
-      { value: 1591160400000, isOtherMonth: true },
-      { value: 1591246800000, isOtherMonth: true },
-      { value: 1591333200000, isOtherMonth: true },
-      { value: 1591419600000, isOtherMonth: true },
-    ],
-    id: 5,
-  },
-];
